// src/models/apiTaxComplianceTvrPeriodListParams.ts
/**
 * Generated by orval
 * Do not edit manually.
 */
import type { ApiTaxComplianceTvrPeriodListWorkflowStatus } from './apiTaxComplianceTvrPeriodListWorkflowStatus'

export type ApiTaxComplianceTvrPeriodListParams = {
  client?: number
  /**
   * Multiple values may be separated by commas.
   */
  client__in?: number[]
  created_at?: string
  created_at__gte?: string
  created_at__lte?: string
  id?: number
  /**
   * Multiple values may be separated by commas.
   */
  id__in?: number[]
  /**
   * Which field to use when ordering the results.
   */
  ordering?: string
  /**
   * A page number within the paginated result set.
   */
  page?: number
  /**
   * Number of results to return per page.
   */
  page_size?: number
  period_month?: number
  period_year?: number
  /**
   * * `DRAFT` - Draft
   * `PREPARED` - Prepared
   * `REVIEW_COMMENTS` - Review Comments
   * `PUBLISHED` - Published
   * `FUNDING_RECEIVED` - Funding Received
   */
  workflow_status?: ApiTaxComplianceTvrPeriodListWorkflowStatus
  /**
   * Multiple values may be separated by commas.
   */
  workflow_status__in?: string[]
}