// src/models/tVRPeriod.ts
/**
 * Generated by orval
 * Do not edit manually.
 */
import type { SimpleClient } from './simpleClient'
import type { WorkflowStatusEnum } from './workflowStatusEnum'

export interface TVRPeriod {
  readonly id: number
  readonly client: SimpleClient
  client_id: number
  /**
   * @minimum 1
   * @maximum 12
   */
  period_month: number
  /**
   * @minimum 2000
   * @maximum 3000
   */
  period_year: number
  /** @nullable */
  backup_file?: string | null
  workflow_status?: WorkflowStatusEnum
  is_expired?: boolean
}